move main code to a trait to use for both printer classes... this breaks history, but allows for a single place to change things
